commit_message_validator: homogeinize case in footers
Homogeinize case in commit message footers. The logic used is that
the '-by' part (i.e, the last part) is always lower case.
In the case of 'Change-Id', I left that untouched since may have impact
in how gerrit and other tools process changes. Same for 'Depends-On', which
integrates with Zuul.

While at it, add the missing 'Requested-by' line.

Tests are updated as well.

Change-Id: Iefc7f9811266cc629f5c43a008b612640a3d8030
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>