Add GitHubMessageValidator and add logic to switch the MessageValidator
Add GitHubMessageValidator and add logic
to detect the remote repo and switch the MessageValidator
to the appropriate MessageValidator, e.g repo with remote
GitHub will be checked with GitHubMessageValidator, and
repo with remote Gerrit will be checked with GerritMessageValidator.

This commit also modifies the way how the tests work,
the `tests/data` directory has subdirectories
that's where the tests for each MessageValidator
is put in.

Bug: T179905
Change-Id: I990c6d7ac313292e929b1c1aba2e6b9e19fb3e7d