Add support for linting multiple commit messages as in a merge request
With support for GitLab it becomes interesting and useful to support
linting of all commit messages in a merge request. CI systems for both
GitLab and GitHub expose envvars that we can use to determine where to
stop linting. These envvars will populate a new `--merge-target` cli
option when present.

* Add `--merge-target` option for `validate` command.
* Hoist output header and footer up to lint.validate()
* Extract lint.validate_single_commit from lint.validate()
* Output "Linting $SHORT_HASH: $SUBJECT" for each commit when there are
  multiple.
* Remove header and footer text from test expectations.

Bug: T339307