ci: Use CI_MERGE_REQUEST_REF_PATH as source branch/ref
When a pipeline runs for a merge request created from a fork, the origin
remote is that of the target repo and there is no remote set up for the
source repo. Because of this, `$CI_MERGE_REQUEST_SOURCE_BRANCH_NAME` is
not likely to exist, and if it happens to it will not contain the MR
commits.

Use `$CI_MERGE_REQUEST_REF_PATH` instead which should always exist in
the target repo and contain the MR commits.